feat: cria BuscarEventoUseCase +  BuscarEventoUseCaseImpl e configura como Bean

+++ src/main/java/dev/alexandraemmeline/EventClean/Core/UseCases/BuscarEventoUseCase.java
@@ -4,6 +4,6 @@ import dev.alexandraemmeline.EventClean.Core.Domains.EventoDomain;
 
 public interface BuscarEventoUseCase {
 
-    EventoDomain execute(Integer id);
+    EventoDomain execute(Long id);
 
 }
